delete button works but it thinks that database is null

+++ app/src/main/java/uk/ac/aston/cs3mdd/mobiledesignproject/ui/Module/data/ModuleListAdapter.java
@@ -125,7 +125,21 @@ public class ModuleListAdapter extends RecyclerView.Adapter<ModuleListAdapter.Mo
             ButtonDeleteModule.setOnClickListener(new View.OnClickListener() {
                 @Override
                 public void onClick(View v) {
-//                    DeleteModuleInBackground(modules);
+                    //Calling all fields
+                    String moduleName = modules.getModuleName();
+                    String moduleCode = modules.getModuleCode();
+
+                    String assignmentDate = modules.getAssignmentDate();
+                    String assignmentdue = modules.getAssignmentdue();
+                    String assignmentName = modules.getAssignmentName();
+
+                    String examdate = modules.getExamdate();
+                    String examdue = modules.getExamdue();
+                    String examName = modules.getExamName();
+
+                    // Deletes the Module from the database using the delete background task
+                    Module DeleteModule = new Module(moduleName, moduleCode, assignmentName, assignmentdue, assignmentDate, examName, examdue, examdate);
+                    DeleteModuleInBackground(DeleteModule);
 
                     Log.i("MS", "Deleted");
                 }
@@ -143,25 +157,28 @@ public class ModuleListAdapter extends RecyclerView.Adapter<ModuleListAdapter.Mo
         }
 
         public void DeleteModuleInBackground(Module module) {
-
             ExecutorService executorService = Executors.newSingleThreadExecutor();
-
             Handler handler = new Handler(Looper.getMainLooper());
 
             executorService.execute(new Runnable() {
                 @Override
                 public void run() {
-                    //background task
-                    moduleDB.getModuleDAO().deleteModule(module);
-                    //on finish task
-                    handler.post(new Runnable() {
-                        @Override
-                        public void run() {
+                    // Check if moduleDB is not null before accessing getModuleDAO()
+                    if (moduleDB != null) {
+                        // Background task
+                        moduleDB.getModuleDAO().deleteModule(module);
 
-                            Toast.makeText(ButtonDeleteModule.getContext(), "Deleted From DataBase", Toast.LENGTH_LONG).show();
-
-                        }
-                    });
+                        // On finish task
+                        handler.post(new Runnable() {
+                            @Override
+                            public void run() {
+                                Toast.makeText(ButtonDeleteModule.getContext(), "Deleted From DataBase", Toast.LENGTH_LONG).show();
+                            }
+                        });
+                    } else {
+                        // Log an error or handle the situation where moduleDB is null
+                        Log.e("DeleteModuleInBackground", "ModuleDatabase is null");
+                    }
                 }
             });
         }
